1.2.0

+++ coppafisher/custom_alignment/base.py
@@ -176,13 +176,13 @@ def fuse_custom_and_dapi(nb: Notebook, extract_dir: str, channel: int) -> np.nda
     # The DAPI stitch results are taken from the notebook. This is important so that the images are aligned with the
     # exported spot positions.
     dapi_fused_image = background.generate_global_image(
-        dapi_images, tile_indices, nbp_basic, nb.stitch, np.float32, silent=False
+        dapi_images, tile_indices, nbp_basic, nb.stitch, np.uint16, silent=False
     )
 
     nbp_stitch = NotebookPage("stitch", {"stitch": {"expected_overlap": expected_overlap}})
     nbp_stitch.tile_origin = tile_origins_custom
     custom_fused_image = background.generate_global_image(
-        custom_images, tile_indices, nbp_basic, nbp_stitch, np.float32, silent=False
+        custom_images, tile_indices, nbp_basic, nbp_stitch, np.uint16, silent=False
     )
 
     # The custom image is cropped/padded with zeros to share the same position and shape of the DAPI fused image.
